==> notebooks/05 - orchestration.ipynb <==
# Monitoring ML Training Pipeline: Orchestration

### Recap:
- **Goal**:
    - Building a classification model for loan eligibility that predicts whether a loan is to be given or refused.
    - Monitor data drifts and model drifts.
- **Download Raw Data**: `raw/12196ecaa65e4831987aee4bfced5f60_2015-01-01_2015-05-31.csv`
- **Preprocessed the Data Into**:
    - Training dataset: `preprocessed/12196ecaa65e4831987aee4bfced5f60.csv`
    - Test dataset: `preprocessed/12196ecaa65e4831987aee4bfced5f60.csv`

- **Trained and Deployed the Model**:
    - JobID: 12196ecaa65e4831987aee4bfced5f60
    - Missing values: 12196ecaa65e4831987aee4bfced5f60_missing_values_model.pkl
    - Purpose to Integer: 12196ecaa65e4831987aee4bfced5f60_purpose_to_int_model.json
    - Prediction model: 12196ecaa65e4831987aee4bfced5f60_rf.pkl

    ---

### Next Steps:
- **Orchestration of the Monitoring Steps and Process**
    - [I. Install Docker](https://docs.docker.com/get-docker/)
    - [II. Deploy a Docker Container for Airflow](#ii-deploy-a-docker-container-for-airflow)
    - [III. Prepare Slack for Integration with Airflow](#iii-prepare-slack-for-integration-with-airflow)
    - [IV. Create Training DAG](#iv-create-training-dag)

    ---

### II. Deploy a Docker Container for Airflow
#### Folder Structuring
```
├── dags
│   ├── data
│   │   ├── preprocessed
│   │   └── raw
│   ├── models
│   ├── results
│   └── src
docker-compose.yaml
├── jobs
├── logs
└── plugins
```

    ---

#### Services
- **UI**

    ---

### III. Prepare Slack for Integration with Airflow
- Create a Slack app and generate a token: [Slack API Apps](https://api.slack.com/apps)
- Create a Slack connection in Airflow
- Test Airflow-Slack integration

    ---

### IV. Create Training DAG


# Import necessary libraries and modules
import datetime
import sys
import os
import json
import re
import pickle
import traceback
import pandas as pd
import numpy as np

##### queries.py #####

# SQL query to create the mljob table if it doesn't exist
CREATE_TABLE_ML_JOB = """
create table if not exists mljob (
    id serial primary key,
    job_id varchar(36) not null,
    job_type varchar(36), -- training, inference
    job_date date, 
    stage varchar(36) not null, -- etl, preprocess, training, deploy, predicting
    status varchar(36) not null, -- started, pass, failed
    message text not null, -- e.g: exceptions
    created_at timestamp not null default now()
);
"""

# SQL query to insert log activity into the mljob table
LOG_ACTIVITY = """
    insert into mljob (
        job_id,
        job_type,
        job_date,
        stage,
        status,
        message
    ) values ('{job_id}', '{job_type}', '{job_date}', '{stage}', '{status}', '{message}')
"""

# SQL query to get the latest deployed job ID
GET_LATEST_DEPLOYED_JOB_ID = """
    select job_id from mljob
    where status = '{status}' and job_type = 'training' and stage = 'deploy'
    order by created_at desc
    limit 1
"""

##### helpers.py #####

# ---> job handlers

# Function to generate a random UUID
def generate_uuid() -> str:
    """
    Generate a random UUID.
    :return: str
    """
    return str(uuid.uuid4()).replace("-", "")

# ---> db handlers

# Function to log activity of a job
def log_activity(job_id:str, job_type:str, stage:str, status:str, message:str, job_date:datetime.date=None):
    """
    Logs the activity of a job.
    :param job_id: str
    :param job_type: str
    :param stage: str
    :param status: str
    :param message: str
    :param job_date: datetime.date
    :return: None
    """
    assert stage in config.STAGES, f"[ERROR] Stage `{stage}` is not valid! Choose from {config.STAGES}"
    assert status in config.STATUS, f"[ERROR] Status `{status}` is not valid! Choose from {config.STATUS}"
    assert job_type in config.JOB_TYPES, f"[ERROR] Job type `{job_type}` is not valid! Choose from {config.JOB_TYPES}"
    message = message.replace("'", "\\")
    engine.execute(text(queries.LOG_ACTIVITY.format(job_id=str(job_id), job_type=job_type, stage=str(stage), status=str(status), message=message, job_date=job_date)).execution_options(autocommit=True))
    print(f"[INFO] Job {job_id} logged as {job_type}::{stage}::{status}::{message}")

# Function to get the latest deployed job ID
def get_latest_deployed_job_id(status:str="pass") -> str:
    """
    Get the latest deployed job id by looking for the latest of all jobs with stage `deploy` and the specified status.
    :param status: str
    :return: str
    """
    try:
        return json.load(open(os.path.join(config.PATH_DIR_MODELS, "deploy_report.json"))).get("job_id")
    except Exception as e:
        assert status in config.STATUS, f"[ERROR] Status `{status}` is not valid! Choose from {config.STATUS}"
        query = text(queries.GET_LATEST_DEPLOYED_JOB_ID.format(status=status))
        r = pd.read_sql(query, engine)
        if r.shape(0) == 0:
            return None
        return str(r['job_id'].values[0])

# ---> files handlers

# Function to load a dataset from a file path
def load_dataset(path:str) -> pd.DataFrame:
    """
    Load data set.
    :param path: str
    :return: DataFrame
    """
    return pd.read_csv(path)

# Function to locate the raw data file for a given job ID
def locate_raw_data_filename(job_id:str) -> str:
    """
    Locate the raw data file.
    :param job_id: str
    :return: str
    """
    files = glob(os.path.join(config.PATH_DIR_DATA, "raw", f"{job_id}_*.csv"))
    if len(files) == 0:
        print(f"[WARNING] No raw data file found for job_id : {job_id}.")
        return None
    return files[0]

# Function to locate preprocessed data file(s) for a given job ID
def locate_preprocessed_filenames(job_id:str) -> dict:
    """
    Locate the preprocessed data files.
    :param job_id: str
    :return: dict
    """
    files = sorted(glob(os.path.join(config.PATH_DIR_DATA, "preprocessed", f"{job_id}_*.csv")))
    if len(files) == 0:
        raise(Exception(f"No preprocessed data file found for job_id : {job_id}."))
    elif len(files) > 2:
        raise(Exception(f"More than one preprocessed data file found for job_id : {job_id} ->\n{files}"))
    elif len(files) == 1:
        training_filename = None
        inference_filename = list(filter(lambda x: "inference" in x, files))[0]
        return training_filename, inference_filename
    else:
        training_filename = list(filter(lambda x: "training" in x, files))[0]
        inference_filename = list(filter(lambda x: "inference" in x, files))[0]
        return training_filename, inference_filename

# Function to get the type of a model
def get_model_type(job_id:str) -> str:
    """
    Get the type of a model.
    :param job_id: str
    :return: str
    """
    report_filename = os.path.join(config.PATH_DIR_MODELS, f"{job_id}_train_report.json")
    return json.load(open(report_filename, "r"))["final_model"]

# Function to load a model from a pickle file
def load_model_from_pickle(model_name: str):
    """
    Load a pickle model.
    :param model_name: str
    :return: model
    """
    with open(os.path.join(config.PATH_DIR_MODELS, model_name+".pkl"), "rb") as f:
        return pickle.load(f)


---
